v4.0.0. This is a complete rewrite of the core captain parsing engine. The previous code had gotten too brittle and there were bugs that were going to be too hard to fix and decisions that I no longer liked.
The interface has been changed to work like endpoints controllers, this makes captain a perfect complement to endpoints.

Fixes #59, dicts passed to captain.io.Output.table() the keys will be column 1 and the values will be column 2, unless all the values are lists, in which case the keys will be the header of each column, and the value will be the rows in each column.

Fixes #57, CAPTAIN_QUIET_DEFAULT is now set to empty string.

Fixes #56, this is actually the issue that prompted the rewrite, I just couldn't make this work with the 3.0.0 branch of the code. Now args and kwargs passed to the handle method will be correctly picked up and passed to the handle method.

Fixes #55, you can have a Default command and subcommands and the Default command will be called when you don't pass in a subcommand. The biggest limitation of this is your Default.handle() method cannot take positional arguments, because argparse will always consider the first positional argument the subcommand.

Fixes #46, this is adjacent to #55.

Closes #45, I'm not going to do this interface idea because I've gone a different direction.

The code is still rough and needs to be cleaned up but it's working and most of the features of the previous iterations are ported over, I did remove some features I never really used.